Changelist Sort: * Remove changelist_sort/workspace/xml_reader.py * Remove changelist_sort/workspace/workspace_tree.py * Move changelist_sort/workspace/developer_file_pattern.py to changelist_sort/sorting/developer_file_pattern.py * Move changelist_sort/workspace/developer_changelist.py to changelist_sort/sorting/developer_changelist.py * Remove changelist_sort/workspace/__init__.py * Remove changelist_sort/input/file_validation.py
